requirements.txt
fastapi
uvicorn
SQLAlchemy
pydantic
passlib[bcrypt]
python-dotenv
python-jose[cryptography]
pytest
